split the app back up into well named stacks
sqs/sns live in wires-feeds, everything downstream in newswires
